Make the Monitor send heartbeats to ensure socket is still viable.
Monitor now sends heartbeats (noop:0 commands) periodically. Default interval between heartbeats is 10 seconds. Heartbeat response (normally 'ok\n') is expected within a third of the heartbeat interval. If a timely heartbeat response isn't received then the connection is closed, status set to stopped, and the retryCallback is called. (retryCallback is passed to the monitor at constructor time.)

First heartbeat is scheduled if the monitor is successfully started.
Subsequent heartbeats are scheduled every time a heartbeat is sent.
Heartbeats are only sent (and the next one scheduled) if the monitor status is still 'started'
Every time a heartbeat is sent, we also schedule the check to verify that a timely response has been received.

The code which deals with the heartbeat response has been added to the _handleResponse function

Changed logging levels a little

Finally, removed call to remoteSecondary.isAvailable() as it is superfluous - basically all it does is open a socket to the secondary, which is what the Monitor is going to do anyway. No point in checking if we can open a socket in order to then ... open a socket.

Added unit tests to cover the heartbeat behaviour.